Retarget passkey rp id regression test to getSecurityKeyHostname
The rp-id host selection moved from me-two-step.ts to getSecurityKeyHostname in the dashboard utils when #112611 fixed the passkey login break. Retarget the regression test to that function: it resolves wordpress.com and its subdomains to the canonical default and keeps other hosts on their own hostname.
